+++ src/components/PlaceMap/PlaceMap.tsx
@@ -1,34 +1,9 @@
-import {MutableRefObject, useState, useEffect, useRef} from 'react';
-import leaflet, {LayerGroup, Map} from 'leaflet';
+import {useEffect, useRef} from 'react';
+import leaflet, {LayerGroup} from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 
+import {useMap} from '../../hooks/useMap';
 import {MapProps} from '../../types/types';
-
-const useMap = (mapRef: MutableRefObject<HTMLElement | null>): Map | null => {
-  const [map, setMap] = useState<Map | null>(null);
-  const isRenderedRef = useRef(false);
-
-  useEffect(() => {
-    if (mapRef.current !== null && !isRenderedRef.current) {
-      const instance: Map = leaflet.map(mapRef.current);
-
-      leaflet
-        .tileLayer(
-          'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
-          {
-            attribution:
-              '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
-          }
-        )
-        .addTo(instance);
-
-      setMap(instance);
-      isRenderedRef.current = true;
-    }
-  }, [mapRef]);
-
-  return map;
-};
 
 const defaultIcon = leaflet.icon({
   iconUrl: './img/pin.svg',
