+++ src/hooks/useMap.ts
@@ -1,0 +1,28 @@
+import {MutableRefObject, useState, useEffect, useRef} from 'react';
+import leaflet, { Map} from 'leaflet';
+
+export const useMap = (mapRef: MutableRefObject<HTMLElement | null>): Map | null => {
+  const [map, setMap] = useState<Map | null>(null);
+  const isRenderedRef = useRef(false);
+
+  useEffect(() => {
+    if (mapRef.current !== null && !isRenderedRef.current) {
+      const instance: Map = leaflet.map(mapRef.current);
+
+      leaflet
+        .tileLayer(
+          'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
+          {
+            attribution:
+              '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
+          }
+        )
+        .addTo(instance);
+
+      setMap(instance);
+      isRenderedRef.current = true;
+    }
+  }, [mapRef]);
+
+  return map;
+};
